+++ components/compositing/Cargo.toml
@@ -22,8 +22,8 @@ path = "../msg"
 [dependencies.net]
 path = "../net"
 
-[dependencies.profile]
-path = "../profile"
+[dependencies.profile_traits]
+path = "../profile_traits"
 
 [dependencies.net_traits]
 path = "../net_traits"
