+++ components/compositing/compositor.rs
@@ -32,8 +32,8 @@ use msg::constellation_msg::{ConstellationChan, NavigationDirection};
 use msg::constellation_msg::{Key, KeyModifiers, KeyState, LoadData};
 use msg::constellation_msg::{PipelineId, WindowSizeData};
 use png;
-use profile::mem;
-use profile::time::{self, ProfilerCategory, profile};
+use profile_traits::mem;
+use profile_traits::time::{self, ProfilerCategory, profile};
 use script_traits::{ConstellationControlMsg, ScriptControlChan};
 use std::cmp;
 use std::collections::HashMap;
