+++ components/compositing/constellation.rs
@@ -23,8 +23,8 @@ use msg::constellation_msg::{self, ConstellationChan, Failure};
 use net_traits::{self, ResourceTask};
 use net_traits::image_cache_task::ImageCacheTask;
 use net_traits::storage_task::{StorageTask, StorageTaskMsg};
-use profile::mem;
-use profile::time;
+use profile_traits::mem;
+use profile_traits::time;
 use script_traits::{CompositorEvent, ConstellationControlMsg};
 use script_traits::{ScriptControlChan, ScriptTaskFactory};
 use std::borrow::ToOwned;
