+++ components/compositing/pipeline.rs
@@ -16,8 +16,8 @@ use gfx::font_cache_task::FontCacheTask;
 use layers::geometry::DevicePixel;
 use msg::constellation_msg::{ConstellationChan, Failure, FrameId, PipelineId, SubpageId};
 use msg::constellation_msg::{LoadData, WindowSizeData, PipelineExitType, MozBrowserEvent};
-use profile::mem;
-use profile::time;
+use profile_traits::mem;
+use profile_traits::time;
 use net_traits::ResourceTask;
 use net_traits::image_cache_task::ImageCacheTask;
 use net_traits::storage_task::StorageTask;
