+++ components/gfx/Cargo.toml
@@ -20,8 +20,8 @@ path = "../util"
 [dependencies.msg]
 path = "../msg"
 
-[dependencies.profile]
-path = "../profile"
+[dependencies.profile_traits]
+path = "../profile_traits"
 
 [dependencies.style]
 path = "../style"
