+++ components/layout/Cargo.toml
@@ -37,8 +37,8 @@ path = "../plugins"
 [dependencies.net_traits]
 path = "../net_traits"
 
-[dependencies.profile]
-path = "../profile"
+[dependencies.profile_traits]
+path = "../profile_traits"
 
 [dependencies.util]
 path = "../util"
