+++ components/layout/layout_task.rs
@@ -42,9 +42,9 @@ use log;
 use msg::compositor_msg::ScrollPolicy;
 use msg::constellation_msg::Msg as ConstellationMsg;
 use msg::constellation_msg::{ConstellationChan, Failure, PipelineExitType, PipelineId};
-use profile::mem::{self, Report, ReportsChan};
-use profile::time::{self, ProfilerMetadata, profile};
-use profile::time::{TimerMetadataFrameType, TimerMetadataReflowType};
+use profile_traits::mem::{self, Report, ReportsChan};
+use profile_traits::time::{self, ProfilerMetadata, profile};
+use profile_traits::time::{TimerMetadataFrameType, TimerMetadataReflowType};
 use net_traits::{load_bytes_iter, ResourceTask};
 use net_traits::image_cache_task::{ImageCacheTask, ImageCacheResult, ImageCacheChan};
 use script::dom::bindings::js::LayoutJS;
