+++ components/layout_traits/Cargo.toml
@@ -19,8 +19,8 @@ path = "../msg"
 [dependencies.net_traits]
 path = "../net_traits"
 
-[dependencies.profile]
-path = "../profile"
+[dependencies.profile_traits]
+path = "../profile_traits"
 
 [dependencies.util]
 path = "../util"
