+++ components/layout_traits/lib.rs
@@ -5,7 +5,7 @@
 extern crate gfx;
 extern crate script_traits;
 extern crate msg;
-extern crate profile;
+extern crate profile_traits;
 extern crate net_traits;
 extern crate url;
 extern crate util;
@@ -18,8 +18,8 @@ extern crate util;
 use gfx::font_cache_task::FontCacheTask;
 use gfx::paint_task::PaintChan;
 use msg::constellation_msg::{ConstellationChan, Failure, PipelineId, PipelineExitType};
-use profile::mem;
-use profile::time;
+use profile_traits::mem;
+use profile_traits::time;
 use net_traits::ResourceTask;
 use net_traits::image_cache_task::ImageCacheTask;
 use url::Url;
