+++ components/net/Cargo.toml
@@ -7,9 +7,6 @@ authors = ["The Servo Project Developers"]
 name = "net"
 path = "lib.rs"
 
-[dependencies.profile]
-path = "../profile"
-
 [dependencies.net_traits]
 path = "../net_traits"
