+++ components/net/lib.rs
@@ -27,7 +27,6 @@ extern crate png;
 #[macro_use]
 extern crate log;
 extern crate openssl;
-extern crate profile;
 extern crate "rustc-serialize" as rustc_serialize;
 extern crate util;
 extern crate time;
