+++ components/net_traits/Cargo.toml
@@ -13,9 +13,6 @@ git = "https://github.com/servo/rust-geom"
 [dependencies.png]
 git = "https://github.com/servo/rust-png"
 
-[dependencies.profile]
-path = "../profile"
-
 [dependencies.util]
 path = "../util"
