+++ components/profile/lib.rs
@@ -16,13 +16,14 @@
 
 extern crate collections;
 extern crate libc;
+#[macro_use]
+extern crate profile_traits;
 #[cfg(target_os="linux")]
 extern crate regex;
 #[cfg(target_os="macos")]
 extern crate task_info;
 extern crate "time" as std_time;
 extern crate util;
-extern crate url;
 
 pub mod mem;
 pub mod time;
