+++ tests/unit/net/Cargo.toml
@@ -14,9 +14,6 @@ path = "../../../components/net"
 [dependencies.net_traits]
 path = "../../../components/net_traits"
 
-[dependencies.profile]
-path = "../../../components/profile"
-
 [dependencies.util]
 path = "../../../components/util"
